+++ src/Common/ObjectArray.php
@@ -10,6 +10,8 @@ namespace andrefelipe\Orchestrate\Common;
  */
 class ObjectArray implements \ArrayAccess, \Countable, ToJsonInterface
 {
+    use ObjectArrayTrait;
+
     /**
      * @param array $values Values to set to the object on construct.
      */
@@ -20,55 +22,6 @@ class ObjectArray implements \ArrayAccess, \Countable, ToJsonInterface
                 $this->{(string) $key} = $value;
             }
         }
-    }
-
-    // public function __toString()
-    // {
-    //     echo $this->toJson();
-    // }
-
-    public function __get($key)
-    {
-        return isset($this->{$key}) ? $this->{$key} : null;
-    }
-
-    public function __set($key, $value)
-    {
-        if (is_array($value)) {
-            $this->{$key} = new self($value);
-        } else {
-            $this->{$key} = $value;
-        }
-    }
-
-    public function __unset($key)
-    {
-        return $this->{$key} = null;
-    }
-
-    public function offsetGet($offset)
-    {
-        return $this->{$offset};
-    }
-
-    public function offsetSet($offset, $value)
-    {
-        $this->{(string) $offset} = $value;
-    }
-
-    public function offsetUnset($offset)
-    {
-        $this->{$offset} = null;
-    }
-
-    public function offsetExists($offset)
-    {
-        return isset($this->{$offset});
-    }
-
-    public function count()
-    {
-        return count(get_object_vars($this));
     }
 
     /**
@@ -94,58 +47,5 @@ class ObjectArray implements \ArrayAccess, \Countable, ToJsonInterface
             }
         }
         return $this;
-    }
-
-    /**
-     * Helper method to merge instances.
-     *
-     * @param object $object
-     * @param object $instance = null
-     *
-     * @return ObjectArray this
-     */
-    private function _mergeObject($object, $instance = null)
-    {
-        if (!is_object($instance)) {
-            $instance = $this;
-        }
-        foreach (get_object_vars($object) as $key => $value) {
-            
-            if (isset($instance->{$key}) && is_object($value) && is_object($instance->{$key})) {
-                $this->_mergeObject($value, $instance->{$key});
-            } else {
-                $instance->{$key} = $value;
-            }
-        }
-        return $instance;
-    }
-
-    public function toArray()
-    {
-        $result = [];
-
-        foreach (get_object_vars($this) as $key => $value) {
-            
-            if ($value === null) {
-                continue;
-            }
-
-            if (is_object($value)) {
-                if (method_exists($value, 'toArray')) {
-                    $result[$key] = $value->toArray();
-                } else {
-                    $result[$key] = $value;
-                }
-            } else {
-                $result[$key] = $value;
-            }
-        }
-
-        return $result;
-    }
-
-    public function toJson($options = 0, $depth = 512)
-    {
-        return json_encode($this->toArray(), $options, $depth);
-    }
+    }   
 }