+++ src/Objects/ListInterface.php
@@ -1,0 +1,47 @@
+<?php
+namespace andrefelipe\Orchestrate\Objects;
+
+/**
+ * 
+ */
+interface ListInterface
+{
+    /**
+     * Get the results of the list operation.
+     *
+     * @return ObjectArray
+     */
+    public function getResults();
+
+    /**
+     * Recursively merge one list results into another.
+     * 
+     * @param ListInterface $list
+     */
+    public function mergeResults(ListInterface $list);
+
+    /**
+     * @return int
+     */
+    public function getTotalCount();
+
+    /**
+     * @return string
+     */
+    public function getNextUrl();
+
+    /**
+     * @return string
+     */
+    public function getPrevUrl();
+
+    /**
+     * @return boolean Success of operation.
+     */
+    public function next();
+
+    /**
+     * @return boolean Success of operation.
+     */
+    public function prev();
+}